+++ CHANGES.md
@@ -13,6 +13,8 @@ To be released.
  -  Added *drop.ts* module.
  -  Added `drop()` function.
  -  Added `dropWhile()` function.
+ -  Added *filter.ts* module.
+ -  Added `filter()` function.
  -  Added *fold.ts* module.
  -  Added `fromIterable()` function.
  -  Added `reduce()` function.